fix: change nonce parameters from u8 to u64 to support larger nonce values (#58)
- Fix nonce parameter in TransactionArgs from u8 to u64
- Fix nested_safe_nonce parameter in TransactionArgs from u8 to u64
- Fix nonce parameter in tx_signing_hashes function from u8 to u64

This fixes the issue where nonce values > 255 would cause 'xxx is not in 0..=255' errors.
The u8 type was too restrictive for Safe transaction nonces which can grow much larger.

+++ crates/safe-hash/src/tx_signing.rs
@@ -58,7 +58,7 @@ impl TxInput {
 pub fn tx_signing_hashes(
     tx_data: &TxInput,
     safe_address: Address,
-    nonce: u8,
+    nonce: u64,
     chain_id: ChainId,
     safe_version: SafeWalletVersion,
 ) -> SafeHashes {
